Connect homepage component to store
