+++ ui/src/assets/styler.css
@@ -17,7 +17,7 @@ body{
 .banner{
     position: relative;
     z-index: 99;
-    height: 320px;
+    height: 300px;
     top: 0;
     left: 0;
     width: 100%;
@@ -38,7 +38,7 @@ body{
     font-weight: 100;
     font-family: 'helvetica';
     width: 100%;
-    height: 109px;
+    height: 102px;
     background: #00000014;
     background: linear-gradient(rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.6)); /* Standard syntax (must be last) */
 }
