+++ ui/src/index.js
@@ -3,18 +3,28 @@ import ReactDOM from 'react-dom';
 import './assets/styler.css';
 import './assets/customJS';
 import registerServiceWorker from './registerServiceWorker';
-import {MuiThemeProvider} from "material-ui";
+import {MuiThemeProvider, createMuiTheme} from "material-ui";
 import Homepage from './components/homepage';
+import {Provider} from 'react-redux';
+import store from "./store";
+
+const theme = createMuiTheme();
 
 class App extends Component {
     render() {
         return (
-            <MuiThemeProvider>
-                <Homepage />
+            <MuiThemeProvider theme={theme}>
+                <Homepage/>
             </MuiThemeProvider>
         );
     }
 }
 
-ReactDOM.render(<App />, document.getElementById('root'));
+
+const connectedComponents =
+    <Provider store={store}>
+        <App/>
+    </Provider>;
+
+ReactDOM.render(connectedComponents, document.getElementById('root'));
 registerServiceWorker();
