+++ ui/src/reducers/comments.js
@@ -1,0 +1,3 @@
+const initialState = {};
+
+export default initialState;
