+++ ui/src/reducers/users.js
@@ -1,0 +1,3 @@
+const initialState = {};
+
+export default initialState;
